Adding destructor tests for StreamingAggregation
Destructors weren't called for StreamingAggregation and that was fixed at https://github.com/facebookincubator/velox/pull/2065. Hereby adding the tests
